minor updates/renames
