+++ ray-operator/test/e2ekuberneteswas/kubernetes_was_v1alpha2_test.go
@@ -1,4 +1,4 @@
-package e2enativescheduling
+package e2ekuberneteswas
 
 import (
 	"fmt"
